Centralize js method execution inside its JsExecutor class
