+++ rich-html-editor/src/main/java/com/infomaniak/lib/richhtmleditor/executor/JsExecutor.kt
@@ -1,7 +1,6 @@
 package com.infomaniak.lib.richhtmleditor.executor
 
 import android.webkit.WebView
-import com.infomaniak.lib.richhtmleditor.JsExecutableMethod
 
 internal class JsExecutor(private val webView: WebView) : JsLifecycleAwareExecutor<JsExecutableMethod>() {
