+++ rich-html-editor/src/main/java/com/infomaniak/lib/richhtmleditor/executor/KeyboardOpener.kt
@@ -3,8 +3,6 @@ package com.infomaniak.lib.richhtmleditor.executor
 import android.app.Activity
 import android.view.View
 import android.view.inputmethod.InputMethodManager
-import android.webkit.WebView
-import com.infomaniak.lib.richhtmleditor.JsExecutableMethod
 
 internal class KeyboardOpener(private val view: View) : JsLifecycleAwareExecutor<Unit>() {
     override fun executeImmediately(value: Unit) {
